feat(tollgate): refactoring, detector related structs and functions put together
