+++ internal/service/tollgate/tollgate.go
@@ -1,99 +1,123 @@
 package tollgate
 
 import (
+	"database/sql"
+	"encoding/json"
 	"fmt"
 
-	"context"
+	"github.com/openmarketplaceengine/openmarketplaceengine/internal/service/tollgate/detector"
+
+	"github.com/openmarketplaceengine/openmarketplaceengine/dao"
+	"github.com/openmarketplaceengine/openmarketplaceengine/dom"
 )
 
-// Detector represents a tollgate the Subject passes through.
-// Crossed detects if subject Movement has travelled through the tollgate.
-type Detector interface {
-	DetectCrossing(ctx context.Context, movement *Movement) (*Crossing, error)
+const table = "tollgate"
+
+type BBoxes struct {
+	BBoxes   []*detector.BBox
+	Required int
 }
 
-// Line represents a two points line that Subject crosses.
-// longitude latitude corresponding to linear algebra X Y axis.
-type Line struct {
-	Lon1 float64
-	Lat1 float64
-	Lon2 float64
-	Lat2 float64
+func (b *BBoxes) Scan(data interface{}) (err error) {
+	return json.Unmarshal(data.([]byte), &b)
 }
 
-// BBox is a bounded box, represents area defined by two longitudes and two latitudes
-// left,bottom,right,top - lonMin, latMin, lonMax, latMax
-// Longitude and Latitude correspond to linear algebra X and Y axis.
-type BBox struct {
-	LonMin float64
-	LatMin float64
-	LonMax float64
-	LatMax float64
+type GateLine struct {
+	Line *detector.Line
 }
 
-// Alg represents algorithm.
-type Alg uint8
-
-const (
-	LineAlg Alg = iota
-	BBoxAlg
-	VectorAlg
-)
-
-// Crossing represents detected fact of passing through the tollgate by WorkerID.
-type Crossing struct {
-	TollgateID string
-	WorkerID   string
-	Movement   *Movement
-	Direction  Direction
-	Alg        Alg
+func (g *GateLine) Scan(data interface{}) (err error) {
+	return json.Unmarshal(data.([]byte), &g)
 }
 
-// Movement represents a moving SubjectID from one Location to another.
-type Movement struct {
-	SubjectID string
-	From      *Location
-	To        *Location
+type Tollgate struct {
+	ID       dom.SUID  `db:"id"`
+	Name     string    `db:"name"`
+	BBoxes   *BBoxes   `db:"b_boxes"`
+	GateLine *GateLine `db:"gate_line"`
+	Created  dao.Time  `db:"created"`
+	Updated  dao.Time  `db:"updated"`
 }
 
-// Location is longitude, latitude corresponding to linear algebra X, Y axis.
-type Location struct {
-	Lon float64
-	Lat float64
+func (t *Tollgate) Insert(ctx dom.Context) error {
+	now := dao.Time{}
+	now.Now()
+	now.UTC()
+	t.Created = now
+	exec := dao.Insert(table).
+		Set("id", t.ID).
+		Set("name", t.Name).
+		Set("b_boxes", t.BBoxes).
+		Set("gate_line", t.GateLine).
+		Set("created", t.Created)
+	return dao.ExecTX(ctx, exec)
 }
 
-//Direction to North, South, East or West in form of N, S, E, W, NE, NW, SE, SW.
-type Direction string
+func (t *Tollgate) Update(ctx dom.Context) error {
+	now := dao.Time{}
+	now.Now()
+	now.UTC()
+	t.Updated = now
+	exec := dao.Update(table).
+		Set("name", t.Name).
+		Set("b_boxes", t.BBoxes).
+		Set("gate_line", t.GateLine).
+		Set("updated", t.Updated).
+		Where("id = ?", t.ID)
+	return dao.ExecTX(ctx, exec)
+}
 
-// Direction of movement
-// When moving to North - latitude increases until 90.
-// When moving to South - latitude decreases until -90.
-// When moving to East - longitude increases until 180.
-// When moving to West - longitude decreases until -180.
-// If 90/180 limit crossed it jumps to -90/-180 and vice versa.
-// Movement represents a moving subject
-// returns Direction in form of N, S, E, W, NE, NW, SE, SW.
-func (m *Movement) Direction() Direction {
-	fromX := m.From.Lon
-	fromY := m.From.Lat
-	toX := m.To.Lon
-	toY := m.To.Lat
-
-	var pole string
-	var side string
-	if fromY < toY {
-		pole = "N"
+func CreateIfNotExists(ctx dom.Context, tollgate *Tollgate) error {
+	_, err := QueryOne(ctx, tollgate.ID)
+	if err != nil {
+		if err == sql.ErrNoRows {
+			return tollgate.Insert(ctx)
+		}
+		return fmt.Errorf("CreateIfNotExists, query tollgate error: %w", err)
 	}
-	if fromY > toY {
-		pole = "S"
-	}
+	return nil
+}
 
-	if fromX < toX {
-		side = "E"
+func QueryOne(ctx dom.Context, id dom.SUID) (*Tollgate, error) {
+	var t Tollgate
+	err := dao.From(table).
+		Bind(&t).
+		Where("id = ?", id).
+		QueryOne(ctx)
+	if err != nil {
+		return nil, err
 	}
-	if fromX > toX {
-		side = "W"
-	}
+	return &t, nil
+}
 
-	return Direction(fmt.Sprintf("%s%s", pole, side))
+func QueryAll(ctx dom.Context, limit int) ([]*Tollgate, error) {
+	query := dao.From(table).
+		Select("id").
+		Select("name").
+		Select("b_boxes").
+		Select("gate_line").
+		Select("created").
+		Select("updated").
+		OrderBy("created desc").
+		Limit(limit)
+	ary := make([]*Tollgate, 0, limit)
+	err := query.QueryRows(ctx, func(rows *dao.Rows) error {
+		for rows.Next() {
+			var c Tollgate
+			if err := rows.Scan(&c.ID, &c.Name, &c.BBoxes, &c.GateLine, &c.Created, &c.Updated); err != nil {
+				return err
+			}
+			ary = append(ary, &c)
+		}
+		return nil
+	})
+	if err != nil {
+		return nil, err
+	}
+	return ary, nil
+}
+
+func deleteAll(ctx dom.Context) error {
+	del := dao.Delete(table)
+	return dao.ExecTX(ctx, del)
 }
